refactor(router): isolate session round-robin queue
